COMMERCE-4037 headless-commerce-machine-learning-api buildREST

+++ modules/dxp/apps/commerce/headless/headless-commerce-machine-learning-api/src/main/java/com/liferay/headless/commerce/machine/learning/dto/v1_0/ProductInteractionRecommendation.java
@@ -141,16 +141,16 @@ public class ProductInteractionRecommendation {
 	protected Long productId;
 
 	@Schema(description = "The recommendation rank.")
-	public Long getRank() {
+	public Integer getRank() {
 		return rank;
 	}
 
-	public void setRank(Long rank) {
+	public void setRank(Integer rank) {
 		this.rank = rank;
 	}
 
 	@JsonIgnore
-	public void setRank(UnsafeSupplier<Long, Exception> rankUnsafeSupplier) {
+	public void setRank(UnsafeSupplier<Integer, Exception> rankUnsafeSupplier) {
 		try {
 			rank = rankUnsafeSupplier.get();
 		}
@@ -164,7 +164,7 @@ public class ProductInteractionRecommendation {
 
 	@GraphQLField(description = "The recommendation rank.")
 	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected Long rank;
+	protected Integer rank;
 
 	@Schema(description = "The recommended product identifier.")
 	public Long getRecommendedProductId() {
